requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
openpyxl
requests
openai
pytest
pydantic-settings
pdf2image
pytesseract
Pillow